Implement BlanketOverlay as Renderer superclass
Splits off the update-triggering logic from L.Renderer into a new
L.BlanketOverlay class. L.Renderer now only holds logic for keeping
a list of L.Paths.

Intended for #8608.
